user_module/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
boto3>=1.33.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
jinja2>=3.1.2
openai>=1.3.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
httpx>=0.25.0
google-generativeai>=0.3.0
